requirements.txt
Pillow==5.4.1
pynput==1.4
pyobjc==5.1.2
pyobjc-core==5.1.2
pyobjc-framework-Accounts==5.1.2
pyobjc-framework-AddressBook==5.1.2
pyobjc-framework-AdSupport==5.1.2
pyobjc-framework-AppleScriptKit==5.1.2
pyobjc-framework-AppleScriptObjC==5.1.2
pyobjc-framework-ApplicationServices==5.1.2
pyobjc-framework-Automator==5.1.2
pyobjc-framework-AVFoundation==5.1.2
pyobjc-framework-AVKit==5.1.2
pyobjc-framework-BusinessChat==5.1.2
pyobjc-framework-CalendarStore==5.1.2
pyobjc-framework-CFNetwork==5.1.2
pyobjc-framework-CloudKit==5.1.2
pyobjc-framework-Cocoa==5.1.2
pyobjc-framework-Collaboration==5.1.2
pyobjc-framework-ColorSync==5.1.2
pyobjc-framework-Contacts==5.1.2
pyobjc-framework-ContactsUI==5.1.2
pyobjc-framework-CoreAudio==5.1.2
pyobjc-framework-CoreAudioKit==5.1.2
pyobjc-framework-CoreBluetooth==5.1.2
pyobjc-framework-CoreData==5.1.2
pyobjc-framework-CoreLocation==5.1.2
pyobjc-framework-CoreMedia==5.1.2
pyobjc-framework-CoreMediaIO==5.1.2
pyobjc-framework-CoreML==5.1.2
pyobjc-framework-CoreServices==5.1.2
pyobjc-framework-CoreSpotlight==5.1.2
pyobjc-framework-CoreText==5.1.2
pyobjc-framework-CoreWLAN==5.1.2
pyobjc-framework-CryptoTokenKit==5.1.2
pyobjc-framework-DictionaryServices==5.1.2
pyobjc-framework-DiscRecording==5.1.2
pyobjc-framework-DiscRecordingUI==5.1.2
pyobjc-framework-DiskArbitration==5.1.2
pyobjc-framework-DVDPlayback==5.1.2
pyobjc-framework-EventKit==5.1.2
pyobjc-framework-ExceptionHandling==5.1.2
pyobjc-framework-ExternalAccessory==5.1.2
pyobjc-framework-FinderSync==5.1.2
pyobjc-framework-FSEvents==5.1.2
pyobjc-framework-GameCenter==5.1.2
pyobjc-framework-GameController==5.1.2
pyobjc-framework-GameKit==5.1.2
pyobjc-framework-GameplayKit==5.1.2
pyobjc-framework-ImageCaptureCore==5.1.2
pyobjc-framework-IMServicePlugIn==5.1.2
pyobjc-framework-InputMethodKit==5.1.2
pyobjc-framework-InstallerPlugins==5.1.2
pyobjc-framework-InstantMessage==5.1.2
pyobjc-framework-Intents==5.1.2
pyobjc-framework-IOSurface==5.1.2
pyobjc-framework-iTunesLibrary==5.1.2
pyobjc-framework-LatentSemanticMapping==5.1.2
pyobjc-framework-LaunchServices==5.1.2
pyobjc-framework-libdispatch==5.1.2
pyobjc-framework-LocalAuthentication==5.1.2
pyobjc-framework-MapKit==5.1.2
pyobjc-framework-MediaAccessibility==5.1.2
pyobjc-framework-MediaLibrary==5.1.2
pyobjc-framework-MediaPlayer==5.1.2
pyobjc-framework-MediaToolbox==5.1.2
pyobjc-framework-ModelIO==5.1.2
pyobjc-framework-MultipeerConnectivity==5.1.2
pyobjc-framework-NaturalLanguage==5.1.2
pyobjc-framework-NetFS==5.1.2
pyobjc-framework-Network==5.1.2
pyobjc-framework-NetworkExtension==5.1.2
pyobjc-framework-NotificationCenter==5.1.2
pyobjc-framework-OpenDirectory==5.1.2
pyobjc-framework-OSAKit==5.1.2
pyobjc-framework-Photos==5.1.2
pyobjc-framework-PhotosUI==5.1.2
pyobjc-framework-PreferencePanes==5.1.2
pyobjc-framework-PubSub==5.1.2
pyobjc-framework-QTKit==5.1.2
pyobjc-framework-Quartz==5.1.2
pyobjc-framework-SafariServices==5.1.2
pyobjc-framework-SceneKit==5.1.2
pyobjc-framework-ScreenSaver==5.1.2
pyobjc-framework-ScriptingBridge==5.1.2
pyobjc-framework-SearchKit==5.1.2
pyobjc-framework-Security==5.1.2
pyobjc-framework-SecurityFoundation==5.1.2
pyobjc-framework-SecurityInterface==5.1.2
pyobjc-framework-ServiceManagement==5.1.2
pyobjc-framework-Social==5.1.2
pyobjc-framework-SpriteKit==5.1.2
pyobjc-framework-StoreKit==5.1.2
pyobjc-framework-SyncServices==5.1.2
pyobjc-framework-SystemConfiguration==5.1.2
pyobjc-framework-UserNotifications==5.1.2
pyobjc-framework-VideoSubscriberAccount==5.1.2
pyobjc-framework-VideoToolbox==5.1.2
pyobjc-framework-Vision==5.1.2
pyobjc-framework-WebKit==5.1.2
PyQt5==5.11.3
PyQt5-sip==4.19.13
pytesseract==0.2.6
rumps==0.2.2
six==1.12.0
